edit

+++ AttinyPin.h
@@ -4,19 +4,19 @@
 
 	class AttinyPin {
 		private:
-			int num;       // ピン番号
-			int mode;      // モード
+			byte num;       // ピン番号
+			byte mode;      // モード
 		public:
 			// コンストラクタ
-			AttinyPin(int p);
-			void begin(int mode);
-			int getNum(); 
+			AttinyPin(byte p);
+			void begin(byte mode);
+			byte getNum(); 
 	};
 
 	/**
 	 * コンストラクタ
 	 */
-	AttinyPin::AttinyPin(int p){
+	AttinyPin::AttinyPin(byte p){
 		this->num = p;
 		this->mode = OUTPUT;
 	}
@@ -24,7 +24,7 @@
 	/**
 	 * ピン番号取得
 	 */
-	int AttinyPin::getNum(){
+	byte AttinyPin::getNum(){
 		return this->num;
 	}
 
@@ -33,12 +33,12 @@
 	 *
 	 * @param mode OUTPUT(0x00), INPUT(0x01), INPUT_PULLUP(0x02) (0x03はアナログ入力)
 	 */
-	void AttinyPin::begin(int mode){
+	void AttinyPin::begin(byte mode){
 		if(3 < mode){
 			mode = OUTPUT;
 		}
 		if(mode==OUTPUT || mode==INPUT || mode==INPUT_PULLUP){
-			pinMode(this->num,mode);
+			pinMode(this->num, mode);
 		}
 		this->mode = mode;
 	}
